Support updating touch event handler on reused cached components
With current directive installation scheme, a tap listener would not update when a cached component is reused, e.g. when [this documented hack](https://www.npmjs.com/package/vue2-touch-events#how-to-add-extra-parameters) is used. An example to consider:

```html
        <HeroPortrait
                v-for="hero in heroes"
                v-touch:tap="onPortraitTapHacky(hero)"
        />
```
```js
export default {
  props: {
    heroes: {
      type: Array
    }
  }
  methods: {
    onPortraitTapHacky(hero) {
        return function() {
            alert(hero.name)
        };
    },
  }
}
```
Without the fix, when `heroes` prop is changed and the list of portraits is rebuilt, a cached component is used for each updated portrait, and on that cached component the event handler would not correspond to the current hero bound to that component.

I'm not very experienced with vue.js, so please suggest code improvements.